+++ Data/Align/Key.hs
@@ -1,7 +1,9 @@
 -----------------------------------------------------------------------------
--- | Module     :  Data.Aligned.Key
+-- | Module     :  Data.Align.Key
 --
 -- 'These'-based zipping and unzipping of indexed functors.
+--
+-- @since 0.7.1
 module Data.Align.Key (
     AlignWithKey (..)
     ) where
@@ -13,7 +15,7 @@ import Data.Align
 import Data.These
 
 -- Instances
---import Control.Applicative  (ZipList)
+import Control.Applicative  (ZipList)
 import Data.Hashable        (Hashable)
 import Data.HashMap.Strict  (HashMap)
 import Data.IntMap          (IntMap)
@@ -22,6 +24,8 @@ import Data.Sequence        (Seq)
 import Data.Vector          (Vector)
 
 -- | Keyed version of 'Align'.
+--
+-- @since 0.7.1
 class (Keyed f, Align f) => AlignWithKey f where
     -- | Analogous to @'alignWith'@, but also provides an index.
     alignWithKey :: (Key f -> These a b -> c) -> f a -> f b -> f c
@@ -29,7 +33,9 @@ class (Keyed f, Align f) => AlignWithKey f where
 
 instance AlignWithKey Maybe
 instance AlignWithKey []
---instance AlignWithKey ZipList
+
+-- | @since 0.7.6
+instance AlignWithKey ZipList
 instance AlignWithKey Seq
 instance AlignWithKey IntMap
 instance Ord k => AlignWithKey (Map k)